requirements.txt
chainlit==1.1.402
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dateparser==1.2.0
pandas==2.2.2
SQLAlchemy==2.0.32
openpyxl==3.1.5
